IDD_Final_Project/requirements-updated.txt
Adafruit-Blinka==8.67.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-apds9960==3.1.17
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-mpr121==2.1.25
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-rgb-display==3.14.3
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
lgpio==0.2.2.0
MarkupSafe==3.0.3
paho-mqtt==1.6.1
pillow==12.0.0
pyftdi==0.57.1
pygame==2.6.1
pyserial==3.5
python-engineio==4.12.3
python-socketio==5.14.3
pyusb==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
simple-websocket==1.1.0
smbus2==0.5.0
sparkfun-qwiic-joystick==2.0.0
sparkfun-qwiic-proximity==0.9.0
sparkfun_qwiic_i2c==2.0.0
sysv-ipc==1.1.0
typing_extensions==4.15.0
Werkzeug==3.1.3
wsproto==1.3.1
